challenges/2-palindromos/python/luxu_luciano/requirements.txt
pysimplegui-4.60.5